refactor: update Motion component imports in test files
